1.完善分页查询订单 2.已上菜的订单不允许撤单 3.完善上菜功能

+++ src/main/java/com/scuec/restaurant/service/impl/OrderServiceimpl.java
@@ -4,10 +4,14 @@ package com.scuec.restaurant.service.impl;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.scuec.restaurant.dao.OrderDao;
+import com.scuec.restaurant.dao.OrderdetDao;
 import com.scuec.restaurant.entities.Order;
 import com.scuec.restaurant.service.OrderService;
+import com.scuec.restaurant.service.OrderdetailService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.math.BigDecimal;
 
 
 @Service
@@ -16,15 +20,41 @@ public class OrderServiceimpl implements OrderService {
     @Autowired
     private OrderDao orderDao;
 
+    @Autowired
+    private OrderService orderService;
+
+    @Autowired
+    private OrderdetailService orderdetailService;
+
+    @Autowired
+    private OrderdetDao orderdetDao;
+
 
     @Override
-    public int deleteOrderById(String orderId) {
-        return orderDao.deleteOrder(orderId);
+    public int deleteOrderById(String tableId) {
+        Order order = orderDao.getOrderBytableId(tableId);
+        if (order != null){
+            String orderId = order.getOrderId();
+            orderDao.deleteOrder(orderId);
+            orderdetailService.deleteOrderdetId(orderId);
+            return 0;
+        }else
+            return -1;
     }
 
+//    @Override
+//    public int deleteOrderById(String orderId,String tableId) {
+//        Order order = orderDao.getOrderBytableId(tableId);
+//        if (order == null){
+//            return orderDao.deleteOrder(orderId);
+//        }else
+//            return -1;
+//    }
+
     @Override
-    public int updateOrdersta(String orderId) {
-        return orderDao.updateOrdersta(orderId);
+    public int updateOrdersta(String orderId,String foodId) {
+        orderDao.updateOrdersta(orderId);
+        return orderdetDao.updateOrdermenusta(orderId,foodId);
     }
 
     @Override
@@ -38,14 +68,22 @@ public class OrderServiceimpl implements OrderService {
         return orderDao.getOrderList(new Page<>(currentPage, pageSize),orderStatus);
     }
 
+
     @Override
     public int addOrder(String orderId, String tableNo, double orderPrice, String orderStatus) {
         return orderDao.addOrder(orderId,tableNo,orderPrice,orderStatus);
     }
 
     @Override
-    public Order getOrderById(String orderId) {
-        return orderDao.getOrderById(orderId);
+    public Order getOrderBytableId(String tableId) {
+        return orderDao.getOrderBytableId(tableId);
+    }
+
+    @Override
+    public int addOrdermenu(String orderId,double orderPrice) {
+        double price = orderDao.getOrderPrice(orderId);
+        orderPrice = price + orderPrice;
+        return orderDao.updateOrderByorderId(orderId,orderPrice);
     }
